fix exec signal tests on windows
